Got seeds pw hashed. Homeroutes redirecting appropriately.
